benchmark: compare.go can compare the result of 2 benchmarks
